Make noop a real noop
