+++ programs/bpf/Cargo.toml
@@ -50,6 +50,7 @@ members = [
     "rust/panic",
     "rust/param_passing",
     "rust/param_passing_dep",
+    "rust/sanity",
     "rust/sysval",
 ]
